Refactor: filenames and calls to stay off merge conflict
