+++ src/components/Header.js
@@ -7,12 +7,8 @@ import './App.css';
 export default function Header() {
   return (
     <Navbar bg="white" expand="lg" className="border border-bottom-3">
-      <link
-        href="https://cdnjs.cloudflare.com/ajax/libs/tailwindcss/2.0.2/tailwind.min.css"
-        rel="stylesheet"
-      />
       <Container className="p-3 bg-white">
-        <NavLink to="/" className="h2 text-decoration-none text-dark me-5 fw-bold">
+        <NavLink to="/" className="h2 text-decoration-none text-dark me-5 fw-bold d-flex">
           <img src={logo} alt="logo" className="logo" />
           Space Traveler&apos;s Hub
         </NavLink>
@@ -22,9 +18,9 @@ export default function Header() {
             className="ms-auto m-2 my-lg-0"
             navbarScroll
           >
-            <NavLink to="/" className="h5 me-4 link text-decoration-none">Rockets</NavLink>
-            <NavLink to="missions" className="h5 me-4 link text-decoration-none">Missions</NavLink>
-            <NavLink to="profile" className="h5 me-4 link text-decoration-none">My Profile</NavLink>
+            <NavLink to="/" className="h5 me-4 link text-decoration-none" activeClassName="link-active">Rockets</NavLink>
+            <NavLink to="missions" className="h5 me-4 link text-decoration-none" activeClassName="link-active">Missions</NavLink>
+            <NavLink to="profile" className="h5 me-4 link text-decoration-none" activeClassName="link-active">My Profile</NavLink>
           </Nav>
         </Navbar.Collapse>
       </Container>
