+++ src/redux/configureStore.js
@@ -1,7 +1,7 @@
 import { createStore, combineReducers, applyMiddleware } from 'redux';
 import reduxThunk from 'redux-thunk';
 import logger from 'redux-logger';
-import missionsReducer from './missions';
+import missionsReducer from './missions/mission';
 
 const reducer = combineReducers({
   missionsReducer,
